+++ InputGlyphs/Assets/InputGlyphs/Scripts/Runtime/Display/Sprite/InputGlyphSprite.cs
@@ -136,6 +136,11 @@ namespace InputGlyphs.Display
             }
         }
 
+        public void UpdateGlyphs()
+        {
+            UpdateGlyphs(PlayerInput);
+        }
+
         private void UpdateGlyphs(PlayerInput playerInput)
         {
             if (!playerInput.isActiveAndEnabled)