+++ InputGlyphs/Assets/InputGlyphs/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added sample scene for rebinding.
 ### Fixed
 - Fixed an issue where the UI would not update when rebinding buttons on a cloned PlayerInput.actions.
+- Made the UpdateGlyphs() function of the UI component public.
 
 ## [1.2.4] - 2025-02-08
 ### Fixed